backend/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1

# ML and Computer Vision
ultralytics>=8.2.0
opencv-python-headless==4.10.0.84
numpy==1.24.3
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
aiofiles==23.2.1
httpx==0.25.2
PyYAML==6.0.1

# WebSocket
websockets==12.0

# Task Queue (optional)
celery==5.3.4
flower==2.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
